Fix CI flaky: TestDecodeImage_JP2 を -race build から除外
shard 1 でランダムに fail していた TestDecodeImage_JP2 (#1088) を
解消する。本 test は jpeg2000.Encode 経由で 3rd party lib
`mrjoshuak/go-jpeg2000` v1.2.1 の内部 race を踏むため、Go race
detector に検出されていた。

  WARNING: DATA RACE
    Write at ... by goroutine X:
      go-jpeg2000/internal/entropy.(*T1).EncodeFast5()
    Previous read at ... by goroutine Y:
      go-jpeg2000/internal/entropy.(*T1).TruncationPoints()

`(*encoder).encodeTile.func1` が起動する子 goroutine が同 *T1 の
field を mutex 無しで read/write しており、mk-go 側で fix できない。

## 修正

build tag `//go:build !race` を持つ別 file (service_jp2_test.go) に
TestDecodeImage_JP2 を切り出す。これで:

  - 通常 build (= 開発時 `go test ./...`): JP2 round-trip test 実行
  - `-race` build (= CI test-shards): test を compile から除外

上流 library が race fix を release したら build tag を撤去できる。

mediaproxy package coverage は通常 build 90.9% / `-race` build 90.5%
で 90% 閾値クリア。decodeImage 関数自体は 88.9% (JP2 path 含む) で、
JP2 dispatch path の cover は通常 build 経由で担保される。

## 検証

- go test -count=1 ./internal/core/mediaproxy/...        通常 build OK
- go test -race -count=3 ./internal/core/mediaproxy/...  -race 3 連発 OK
- go test -race -coverprofile=... → mediaproxy 90.5%
- make fmt / make lint clean

Closes #1088
