+++ internal/core/mediaproxy/service_test.go
@@ -1,7 +1,6 @@
 package mediaproxy
 
 import (
-	"bytes"
 	"context"
 	"fmt"
 	"image"
@@ -12,7 +11,6 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	jpeg2000 "github.com/mrjoshuak/go-jpeg2000"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -675,26 +673,10 @@ func TestFetch_PassThrough_MNG(t *testing.T) {
 	assert.Equal(t, mngBytes, body)
 }
 
-// TestDecodeImage_JP2 は #734 で追加した JPEG 2000 pure Go decoder
-// (mrjoshuak/go-jpeg2000) の input 経路を確認する。Encode → Decode の
-// round-trip で bounds が保たれれば OK。標準 image.Decode 経由で動作する。
-func TestDecodeImage_JP2(t *testing.T) {
-	// 2x2 RGB image を JP2 で encode してから decode で読み直す
-	src := image.NewRGBA(image.Rect(0, 0, 2, 2))
-	src.Set(0, 0, color.RGBA{R: 255, A: 255})
-	src.Set(1, 0, color.RGBA{G: 255, A: 255})
-	src.Set(0, 1, color.RGBA{B: 255, A: 255})
-	src.Set(1, 1, color.RGBA{R: 255, G: 255, A: 255})
-
-	var buf bytes.Buffer
-	require.NoError(t, jpeg2000.Encode(&buf, src, nil), "encode JP2")
-	require.NotEmpty(t, buf.Bytes())
-
-	img, err := decodeImage(buf.Bytes(), "image/jp2")
-	require.NoError(t, err)
-	assert.Equal(t, 2, img.Bounds().Dx())
-	assert.Equal(t, 2, img.Bounds().Dy())
-}
+// TestDecodeImage_JP2 は service_jp2_test.go に移動した (#1088)。
+// 3rd party lib `mrjoshuak/go-jpeg2000` 内部の data race を踏むため
+// `-race` build から除外する必要があり、build tag (`//go:build !race`)
+// で separate file 化している。
 
 // TestDecodeImage_TGA は #672 Phase 1 で追加した TGA decoder (blezek/tga)
 // が contentType 経由で manual dispatch されることを確認する。
